ete/app/requirements.txt
Flask==2.2.5
mysqlclient==2.1.1
python-dotenv